requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
python-dotenv==1.1.0
google-generativeai==0.8.5
PyPDF2==3.0.1
docx2txt==0.9
pydantic==2.11.4
requests==2.32.3
pytest==8.3.5
pytest-cov==6.1.1
opentelemetry-api==1.32.1
opentelemetry-sdk==1.33.0
opentelemetry-exporter-otlp==1.33.0
prometheus-client==0.21.1